Fixes #34796 - Fixes for Rails 6.1 upgrade
Changes:
- loading of host extension helper
- Rails 6 'where.not' use NAND instead of NOR [0]

[0] https://github.com/rails/rails/pull/36029
